move TypeScript support to a single plugin
