--- AvalonStudio/AvalonStudio.Debugging.TypeScript/packages.config
@@ -1,4 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="IridiumJS" version="2.10.3" targetFramework="net461" />
-</packages>
