+++ AvalonStudio/AvalonStudio.LanguageSupport.TypeScript/LanguageService/TypeScriptDataAssociation.cs
@@ -4,7 +4,7 @@ using AvalonStudio.TextEditor.Rendering;
 using System;
 using System.Collections.Generic;
 
-namespace AvalonStudio.Languages.TypeScript
+namespace AvalonStudio.LanguageSupport.TypeScript.LanguageService
 {
     internal class TypeScriptDataAssociation
     {
