--- AvalonStudio/AvalonStudio.Languages.TypeScript/packages.config
@@ -1,16 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Avalonia" version="0.4.1-build2590-alpha" targetFramework="net461" />
-  <package id="IridiumJS" version="2.10.3" targetFramework="net461" />
-  <package id="jint" version="2.10.2" targetFramework="net461" />
-  <package id="Newtonsoft.Json" version="9.0.2-beta2" targetFramework="net461" />
-  <package id="Serilog" version="1.5.14" targetFramework="net461" />
-  <package id="Splat" version="1.6.2" targetFramework="net461" />
-  <package id="Sprache" version="2.0.0.51" targetFramework="net461" />
-  <package id="System.Reactive" version="3.0.0" targetFramework="net461" />
-  <package id="System.Reactive.Core" version="3.0.0" targetFramework="net461" />
-  <package id="System.Reactive.Interfaces" version="3.0.0" targetFramework="net461" />
-  <package id="System.Reactive.Linq" version="3.0.0" targetFramework="net461" />
-  <package id="System.Reactive.PlatformServices" version="3.0.0" targetFramework="net461" />
-  <package id="System.Reactive.Windows.Threading" version="3.0.0" targetFramework="net461" />
-</packages>
