--- AvalonStudio/AvalonStudio.Projects.TypeScript/packages.config
@@ -1,4 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Newtonsoft.Json" version="9.0.2-beta2" targetFramework="net461" />
-</packages>
