--- AvalonStudio/AvalonStudio.Toolchains.TypeScript/Utilities/ShellExecuteResult.cs
@@ -1,8 +1,0 @@
-﻿namespace AvalonStudio.Toolchains.TypeScript.Utilities
-{
-    public struct ShellExecuteResult
-    {
-        public int ExitCode { get; set; }
-        public string Output { get; set; }
-    }
-}
