--- AvalonStudio/AvalonStudio.Toolchains.TypeScript/Utilities/ShellExecutorType.cs
@@ -1,9 +1,0 @@
-﻿namespace AvalonStudio.Toolchains.TypeScript.Utilities
-{
-    public enum ShellExecutorType
-    {
-        Generic,
-        Windows,
-        Unix
-    }
-}
